requirements.txt
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
iopath==0.1.10
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lapx==0.5.10.post1
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
plotly==5.24.1
portalocker==2.10.1
proglog==0.1.10
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
-e git+https://github.com/facebookresearch/segment-anything-2.git@c2ec8e14a185632b0a5d8b161928ceb50197eddc#egg=SAM_2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
segment-anything==1.0
six==1.16.0
stack-data==0.6.3
supervision==0.24.0
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.8
ultralytics-thop==2.0.9
urllib3==2.2.3
wcwidth==0.2.13
